# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicación web para Comisión EMPRENDE (Componente Backend)",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://backend-emprende.herokuapp.com/"
        }
    ],
    "paths": {
...
        "/api/v1/emprendimiento": {
            "get": {
                "tags": [
                    "Emprendimientos"
                ],
                "summary": "Mostrar información de emprendimientos",
                "operationId": "mostrarEmprendimientos",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los emprendimientos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/emprendimiento/create": {
            "post": {
                "tags": [
                    "Emprendimientos"
                ],
                "summary": "Crear información de emprendimiento",
                "description": " Crear la información de emprendimiento",
                "operationId": "crearEmprendimiento",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "rol_esfot",
                                    "nombre",
                                    "descripcion"
                                ],
                                "properties": {
                                    "rol_esfot": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "cobertura": {
                                        "type": "string"
                                    },
                                    "pagina_web": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "integer"
                                    },
                                    "whatsapp": {
                                        "type": "string"
                                    },
                                    "facebook": {
                                        "type": "string"
                                    },
                                    "instagram": {
                                        "type": "string"
                                    },
                                    "descuento": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/emprendimiento/{emprendimiento}": {
            "get": {
                "tags": [
                    "Emprendimientos"
                ],
                "summary": "Mostrar información de un emprendimiento en específico",
                "operationId": "mostrarEmprendimiento",
                "parameters": [
                    {
                        "name": "emprendimiento",
                        "in": "path",
                        "description": "id de emprendimiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/emprendimiento/{emprendimiento}/update": {
            "post": {
                "tags": [
                    "Emprendimientos"
                ],
                "summary": "Actualizar la información de un emprendimiento",
                "description": " Actualizar la información de una cárcel",
                "operationId": "updateEmprendimiento",
                "parameters": [
                    {
                        "name": "emprendimiento",
                        "in": "path",
                        "description": "id del emprendimiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "rol_esfot": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "categoria": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "cobertura": {
                                        "type": "string"
                                    },
                                    "pagina_web": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "integer"
                                    },
                                    "whatsapp": {
                                        "type": "string"
                                    },
                                    "facebook": {
                                        "type": "string"
                                    },
                                    "instagram": {
                                        "type": "string"
                                    },
                                    "descuento": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/emprendimiento/{emprendimiento}/destroy": {
            "get": {
                "tags": [
                    "Emprendimientos"
                ],
                "summary": "Eliminar emprendimientos",
                "operationId": "destroyEmprendimiento",
                "parameters": [
                    {
                        "name": "emprendimiento",
                        "in": "path",
                        "description": "id de emprendimiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backend-tesis.test/oauth/authorize",
                        "tokenUrl": "http://backend-tesis.test/oauth/token",
                        "refreshUrl": "http://backend-tesis.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}